Refactor admin dashboard and server-side API integration: Replace client-side data fetching with server-side functions for terms and students, streamline the AdminDashboard component, and implement a new API module for improved data handling. Enhance the dashboard layout and functionality for better user experience.
